ICommentService 와 같은 네이밍은 지양합니다.
